requirements.txt
alethic-ism-core
alethic-ism-db
pyyaml
nats-py
python-dotenv
pydantic
psycopg2
google-generativeai
protobuf
mako
